WEB-451 Console command to delete APKs older than X days
This was done to make more space available on our hard drives and get rid
of old APK files, but still keep the newest APKs available to the users.

Example call:
 >php bin/console catrobat:clean:old-apk 3

This would delete all APKs older than three days. The age of APKs depends
on the time of their last access. (Unix atime)

This command depends on the fact that APKs are saved with their ID
as filename to work correctly.

*What changed?*
  CleanOldApkCommand.php created
  commands.yml file updated.