WEB-422 Follow User
Feature that provides the ability to follow and unfollow users. Including a way to dynamically load follower and follows under the profile and myprofile page.
Send Notification when user follows user and when a user you follow creates a new program.

Added following and follower to the User entity.
Added a follow and unfollow icon for the profile page.
Added a gridview for follower to the profile page.
Added a gridview for follower and following to the myprofile page.
Added Notification FollowNotification and ProgramFollowNotification
Added ProfileLoader and added function to ProfileController to load more profiles with pagination.

new User Fields:
- follower [CollectionArray]
- following [CollectionArray]

new Doctrine fields
FollowNotification:
- follower_id
ProgramFollowNotification:
- program_id

- made php7Iwillsafeyou compatible
- fixed checkstyle
- reduced code redundancy a bit