WEB - Recommender System Diversity
In order to improve the diversity of recommended items, the recommender system has been modified. This changes are part of a master's thesis and an online experiment will be conducted. Therefore, two modifications has been made to the original recommendation algorithm (in RecommenderManager.php). For the online experiment a new entity has been created: UserTestGroups.php. With the help of this entity users can be randomly assigned to different test groups. The entity has been made to not rely on assigning users in groups based on location or language, hence eliminating possible biases. The user_id and group_id and datetime are saved in the database. For the online experiment: In RecommenderController.php users get assigned to one of three groups randomly (with the help of UserTestGroup entity) and then one of three recommendation algorithms is presented (baseline + 2 modified versions). Changes to database schema: new table from UserTestGroup entity.