HOTFIX fixes limit/offset of RecSys and algorithm
This hotfix fixes the issue that only 20 programs are shown in the field of recommended programs on the homepage. The issue was that no values for limit and offset were specified in recommendHomepageProgramsForGuests (RecommenderManager.php, on line 239) and thus the default value of 20 was used. 
Also the formula for the weight decrease has been updated slightly to further decrease the weight of popular programs, as the decrease of the old formula might be too low on second thought.