SHARE 95 keep ascpect ratio in thumbnails
Thumbnails are generated when the user uploads a project / we import a project.
Until now images were just reduced in size (fixed width, height -> square), which resulted in distorted images
Now we crop a square with the correct size using imagicks cropThumbnail function to keep the aspect ratios as they were.